Final Commit
Ready for Review. Renamed the files to their proper names. Added in the event listener to fire when the page is loaded. Added a function the validates the summary. Added a function that sums up every input with a certain class. Added a function that calculates all of the values when an input value is changed. Called in the script function to the index. Added in the regex for the input elements in the html.